Small screen + web fixes + new logo
Small screen factor & splash screen if JS off, logo, loader fix
