+++ docs/JS/core.js
@@ -1,9 +1,19 @@
-function tabchange(id) {
-    $('nav a[href="#' + id + '"]').trigger('click');
+function tabchange(targetId) {
+    $('section').addClass('hide');
+    $('.nav-link').removeClass('active');
+    $(`nav a[tab_Id="${targetId}"]`).addClass('active');
+    $(`#${targetId}`).removeClass('hide');
+    var urlParams = new URLSearchParams(window.location.search);
+    urlParams.set('page', targetId);
+    history.replaceState(null, null, '?' + urlParams.toString())
+    if ($('#menuToggle').find('span').hasClass('fa-times')) {
+        $('#menuToggle').trigger('click');
+    }
 }
-$('#menuToggle').click(function () {
-    $(this).find('span').toggleClass('fa-bars fa-times'); // Toggle between bars and times icons
-});
+
+
+// Toggle between bars and times icons
+$('#menuToggle').click(function () { $(this).find('span').toggleClass('fa-bars fa-times'); });
 
 function animateValue(obj, start, end, duration) {
     return new Promise(resolve => {
@@ -21,4 +31,3 @@ function animateValue(obj, start, end, duration) {
         window.requestAnimationFrame(step);
     });
 }
-
